make pytest.sh work both for unittest and integrationtest

+++ docker/api/pytest.sh
@@ -3,15 +3,22 @@
 PYTESTARGS=""
 
 if [ ! -z "$NO_EQUIPMENTTEST" ] ; then
-	PYTESTARGS="-m 'not equipment'"
+	PYTESTARGS="$PYTESTARGS -m 'not equipment'"
+fi
+
+if [ ! -z "$COVERAGE" ] ; then
+	PYTESTARGS="$PYTESTARGS --cov=cnaas_nms -p no:cacheprovider"
 fi
 
 cd /opt/cnaas/venv/cnaas-nms/src/
 
 source ../../bin/activate
 
-pytest --cov=cnaas_nms -p no:cacheprovider $PYTESTARGS
+pytest $PYTESTARGS
 EXITSTATUS="$?"
-mv .coverage .coverage-pytest
+
+if [ -f ".coverage" ] ; then
+	mv .coverage .coverage-pytest
+fi
 
 exit $EXITSTATUS
